refactor i18n libraries
