+++ frontend/src/translations/zh-hans.global.json
@@ -2,16 +2,23 @@
   "app": {
     "name": "NEAR 浏览器"
   },
-  "home": {
-    "title": {
-      "explore": "探索",
-      "near_blockchain": "<br />NEAR区块链"
+  "page": {
+    "home": {
+      "title": {
+        "explore": "探索",
+        "near_blockchain": "<br />NEAR区块链"
+      }
     }
   },
-  "utils": {
-    "Search": {
-      "title": "搜索",
-      "hint": "搜索账户ID、交易哈希、块哈希、块高度"
+  "component": {
+    "utils": {
+      "Search": {
+        "title": "搜索",
+        "hint": "搜索账户ID、交易哈希、块哈希、块高度"
+      },
+      "Header": {
+        "home": "主页"
+      }
     }
   }
 }
